requirements.txt
flask
spotipy
scikit-learn
python-dotenv